+++ hopnet-storage/src/api.rs
@@ -434,7 +434,7 @@ impl StateReader for NullNet {
     fn placement_inputs(&self) -> Result<crate::traits::PlacementInputs, StorageError> {
         Err(StorageError::Host("null state reader".into()))
     }
-    fn placement_inputs_at(&self, _height: i32) -> Result<crate::traits::PlacementInputs, StorageError> {
+    fn placement_inputs_at(&self, _height: u64) -> Result<crate::traits::PlacementInputs, StorageError> {
         Err(StorageError::Host("null state reader".into()))
     }
     fn fragment_sources(
@@ -1128,7 +1128,7 @@ mod tests {
         }
         fn placement_inputs_at(
             &self,
-            height: i32,
+            height: u64,
         ) -> Result<crate::traits::PlacementInputs, StorageError> {
             Ok(crate::traits::PlacementInputs {
                 height,
